Wrote neato Coq and Agda code ^_^
Code some tactics to demonstrate Coq proofs
then showed their equivalents in Agda.

It was my first time using Agda's “record” syntax
solely for namespace management, as that is the purpose
of the “module” keyword. This was always claimed doable,
but was no necessary to make the comparison with the Coq
code.