thesis_ch_4_1.pdf  ⇐  Working on the github issues (•̀ᴗ•́)و
Progress towards issue #28, namely points i and iii; not yet item ii.

Most of the suggestions of issue #29 have been adopted, including the reduction
of bold. Issue #29 may be ready to be closed. (Sorry for not engaging too much
on the github issue thread!)

Regarding issue #31, section 4.1 has been reduced: The paragraph “Why textual
transformations?” has been discarded. It may reappear in a terser form in an
introduction to reflection, as a new section 5.1 ---perhaps similarly for
issues #32 and #34.  For now, issue #31 may be ready to be closed.

Issue #33 not yet addressed.

Advice of issue #35 followed fully (•̀ᴗ•́)و

Have also progressed toward issue #30. There are likely other places in the
thesis where this is a concern, so the issue should continue to remain open.